Update google regex to fire only when addressed
Addresses Issue #36

Updates the googling regex by adding '(?i)gthx' at the start of the regex which will cause it to only fire when gthx is directly addressed. The '(?i)' portion is an inline modifier to make the regex case insensitive, which will match:
gthx google ...
gthx Google ...
GTHX Google ...
Gthx google ...
etc... 

This may not be strictly needed due to tab completion defaulting to gthx , however it may be found to be necessary to change the regex to have an autofilled nick instead of hardcoding 'gthx' into it in the case of our nick being lost due to netsplit.